requirements.txt
PySide6==6.6.2
pillow==10.2.0
pyinstaller